+++ app/src/main/java/woowacourse/paint/adapter/ColorPaletteViewHolder.kt
@@ -1,8 +1,9 @@
-package woowacourse.paint
+package woowacourse.paint.adapter
 
 import androidx.annotation.ColorRes
 import androidx.core.content.ContextCompat
 import androidx.recyclerview.widget.RecyclerView.ViewHolder
+import woowacourse.paint.ColorPaletteHandler
 import woowacourse.paint.databinding.ItemColorPaletteBinding
 
 class ColorPaletteViewHolder(
