+++ app/src/main/java/woowacourse/paint/drawingboard/Line.kt
@@ -1,4 +1,4 @@
-package woowacourse.paint
+package woowacourse.paint.drawingboard
 
 import android.graphics.Paint
 import android.graphics.Path
